Reorganized concordia.utils
